requirements.txt
# Flask compatible with Python 3.6
Flask==1.1.4

# Feed parsing
feedparser==6.0.10

# HTTP requests
requests==2.31.0

# YAML support
PyYAML==6.0

# HTML parsing
beautifulsoup4==4.12.2

# Domain extraction
tldextract==3.4.5

# APScheduler
apscheduler==3.10.5
